+++ src/main/kotlin/TelegramBotService.kt
@@ -1,0 +1,36 @@
+package org.example
+
+import java.net.URI
+import java.net.URLEncoder
+import java.net.http.HttpClient
+import java.net.http.HttpRequest
+import java.net.http.HttpResponse
+
+class TelegramBotService(private val botToken: String) {
+
+    fun getUpdates(updateId: Int): String {
+        val urlGetUpdates = "$TELEGRAM_BASE_URL/bot$botToken/getUpdates?offset=$updateId"
+        val client: HttpClient = HttpClient.newBuilder().build()
+        val request: HttpRequest = HttpRequest.newBuilder().uri(URI.create(urlGetUpdates)).build()
+        val response: HttpResponse<String> = client.send(request, HttpResponse.BodyHandlers.ofString())
+        return response.body()
+    }
+
+    fun sendMessage(chatId: Long, text: String): String {
+        val urlSendMessage =
+            "$TELEGRAM_BASE_URL/bot$botToken/sendMessage?chat_id=$chatId&text=${
+                URLEncoder.encode(
+                    text,
+                    "UTF-8"
+                )
+            }"
+        val client: HttpClient = HttpClient.newBuilder().build()
+        val request: HttpRequest = HttpRequest.newBuilder().uri(URI.create(urlSendMessage)).build()
+        val response: HttpResponse<String> = client.send(request, HttpResponse.BodyHandlers.ofString())
+        return response.body()
+    }
+
+    companion object {
+        const val TELEGRAM_BASE_URL = "https://api.telegram.org"
+    }
+}
